organize test files
