+++ packages/malloy-filter/src/test/fstring.spec.ts
@@ -7,18 +7,15 @@
 
 import {diff} from 'jest-diff';
 import * as nearley from 'nearley';
-import fstring_grammar from './lib/fexpr_string_parser';
-import {StringClause, NumberClause} from './clause_types';
-import {StringFilterExpression} from './string_filter_expression';
-import fnumber_grammar from './lib/fexpr_number_parser';
-import {NumberFilterExpression} from './number_filter_expression';
+import fstring_grammar from '../lib/fexpr_string_parser';
+import {StringClause} from '../clause_types';
+import {StringFilterExpression} from '../string_filter_expression';
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace jest {
     interface Matchers<R> {
       isStringFilter(expected: StringClause, unparse?: string): R;
-      isNumberFilter(expected: NumberClause, unparse?: string): R;
     }
   }
 }
@@ -49,47 +46,6 @@ expect.extend({
         };
       }
       const serialize_error = diff(src, unparse, {contextLines: 10});
-      return {
-        pass: false,
-        message: () =>
-          `Unparse Error: '${src}' parsed correctly, but unparsed as '${unparse}'\n${serialize_error}`,
-      };
-    }
-    const errTest = diff(
-      {parse: expectedParse},
-      {parse: results[0]},
-      {expand: true}
-    );
-    return {
-      pass: false,
-      message: () => `${src} did not compile correctly\n${errTest}`,
-    };
-  },
-  isNumberFilter(
-    src: string,
-    expectedParse: NumberClause,
-    expectedUnparse?: string
-  ) {
-    const fnumber_parser = new nearley.Parser(
-      nearley.Grammar.fromCompiled(fnumber_grammar)
-    );
-    fnumber_parser.feed(src);
-    const results = fnumber_parser.finish();
-    if (results.length > 1) {
-      return {
-        pass: false,
-        message: () => 'Ambiguous parse, grammar error',
-      };
-    }
-    if (this.equals(expectedParse, results[0])) {
-      const unparse = NumberFilterExpression.unparse(results[0]);
-      if (unparse === (expectedUnparse ?? src)) {
-        return {
-          pass: true,
-          message: () => `${src} parsed and serialized correctly`,
-        };
-      }
-      const serialize_error = diff(src, unparse);
       return {
         pass: false,
         message: () =>
@@ -422,136 +378,4 @@ describe('string filter expressions', () => {
   test.todo(
     'write many malformed expressions and generate reasonable errors for them'
   );
-});
-
-describe('number filter expressions', () => {
-  test('null', () => {
-    expect('null').isNumberFilter({operator: 'null'});
-  });
-  test('not null', () => {
-    expect('not null').isNumberFilter({operator: 'null', not: true});
-  });
-  test('N', () => {
-    expect('5').isNumberFilter({operator: '=', values: ['5']});
-  });
-  test('-N', () => {
-    expect('-5').isNumberFilter({operator: '=', values: ['-5']});
-  });
-  test('N.N', () => {
-    expect('4.2').isNumberFilter({operator: '=', values: ['4.2']});
-  });
-  test('-N.NE+N', () => {
-    expect('-4.2E+3').isNumberFilter({operator: '=', values: ['-4.2E+3']});
-  });
-  test('NEN', () => {
-    expect('4E-2').isNumberFilter({operator: '=', values: ['4E-2']});
-  });
-  test('= N', () => {
-    expect('= 42').isNumberFilter({operator: '=', values: ['42']}, '42');
-  });
-  test('!= N', () => {
-    expect('!= 42').isNumberFilter({operator: '!=', values: ['42']});
-  });
-  test('< N', () => {
-    expect('< 42').isNumberFilter({operator: '<', values: ['42']});
-  });
-  test('> N', () => {
-    expect('> 42').isNumberFilter({operator: '>', values: ['42']});
-  });
-  test('<= N', () => {
-    expect('<= 42').isNumberFilter({operator: '<=', values: ['42']});
-  });
-  test('>= N', () => {
-    expect('>= 42').isNumberFilter({operator: '>=', values: ['42']});
-  });
-  test('(1 to 2)', () => {
-    expect('(1 to 2)').isNumberFilter({
-      operator: 'range',
-      startOperator: '>',
-      startValue: '1',
-      endOperator: '<',
-      endValue: '2',
-    });
-  });
-  test('(1 to 2]', () => {
-    expect('(1 to 2]').isNumberFilter({
-      operator: 'range',
-      startOperator: '>',
-      startValue: '1',
-      endOperator: '<=',
-      endValue: '2',
-    });
-  });
-  test('[1 to 2]', () => {
-    expect('[1 to 2]').isNumberFilter({
-      operator: 'range',
-      startOperator: '>=',
-      startValue: '1',
-      endOperator: '<=',
-      endValue: '2',
-    });
-  });
-  test('[1 to 2)', () => {
-    expect('[1 to 2)').isNumberFilter({
-      operator: 'range',
-      startOperator: '>=',
-      startValue: '1',
-      endOperator: '<',
-      endValue: '2',
-    });
-  });
-  test('1, 2', () => {
-    expect('1, 2').isNumberFilter({operator: '=', values: ['1', '2']});
-  });
-  test('1,2,3', () => {
-    expect('1, 2, 3').isNumberFilter({operator: '=', values: ['1', '2', '3']});
-  });
-  test('1 or 2', () => {
-    expect('1 or 2').isNumberFilter(
-      {operator: '=', values: ['1', '2']},
-      '1, 2'
-    );
-  });
-  test('1 and 2', () => {
-    expect('1 and 2').isNumberFilter({
-      operator: 'and',
-      members: [
-        {operator: '=', values: ['1']},
-        {operator: '=', values: ['2']},
-      ],
-    });
-  });
-  test('1 and 2 and 3', () => {
-    expect('1 and 2 and 3').isNumberFilter({
-      operator: 'and',
-      members: [
-        {operator: '=', values: ['1']},
-        {operator: '=', values: ['2']},
-        {operator: '=', values: ['3']},
-      ],
-    });
-  });
-  test('1,!=2', () => {
-    expect('1, != 2').isNumberFilter({
-      operator: ',',
-      members: [
-        {operator: '=', values: ['1']},
-        {operator: '!=', values: ['2']},
-      ],
-    });
-  });
-  test('just (1)', () => {
-    expect('(1)').isNumberFilter({
-      operator: '()',
-      expr: {operator: '=', values: ['1']},
-    });
-  });
-  test('not (1)', () => {
-    expect('not (1)').isNumberFilter({
-      operator: '()',
-      not: true,
-      expr: {operator: '=', values: ['1']},
-    });
-  });
-  test.todo('all the ways that not can be applied');
 });
